MYNEWT-729 SensorAPI: Send sensor data over OIC
- Move bleprph specific code to the app
